+++ src/main.tsx
@@ -4,6 +4,7 @@ import { ConvexProvider, ConvexReactClient } from "convex/react";
 import { BrowserRouter } from "react-router-dom";
 import App from "./App";
 import { ThemeProvider } from "./context/ThemeContext";
+import { FontProvider } from "./context/FontContext";
 import "./styles/global.css";
 
 const convex = new ConvexReactClient(import.meta.env.VITE_CONVEX_URL as string);
@@ -13,7 +14,9 @@ ReactDOM.createRoot(document.getElementById("root")!).render(
     <ConvexProvider client={convex}>
       <BrowserRouter>
         <ThemeProvider>
-          <App />
+          <FontProvider>
+            <App />
+          </FontProvider>
         </ThemeProvider>
       </BrowserRouter>
     </ConvexProvider>